mysite/requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-debug-toolbar==2.1
django-grappelli==2.13.2
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.4
requests==2.22.0
simplejson==3.17.0
whitenoise==5.0.1

